requirements.txt
streamlit
langchain-groq
langchain-community
langchain-core
pypdf
faiss-cpu
python-dotenv
